+++ Additionals/ChangeLog.md
@@ -6,3 +6,4 @@
 * Detection mode is automatically disabled when then samples tab is deactivated
 * Background color of detection mode changed
 * Clip types updated if clip is removed
+* Scoreboard contents shown in title bar